+++ src/songdata/songdata.h
@@ -1,15 +1,9 @@
 #ifndef _songdata_h
 #define _songdata_h
 
-#include <unistd.h>
-#include <stdlib.h>
-
-typedef struct _dirstack
-{
-	struct _dirstack *prev;
-	char *fullpath;
-	char *filename;
-} dirstack;
+#define L_APPEND    0
+#define L_NEW		1
+#define L_SEARCH	2
 
 typedef struct _songdata_song
 {
@@ -34,6 +28,7 @@ typedef struct _songdata_song
 	struct _songdata_song *prev;
 } songdata_song;
 
+
 typedef struct
 {
 	char *from;
@@ -52,25 +47,16 @@ typedef struct
 #define F_VIRTUAL      	0x01
 } songdata;
 
-#define L_APPEND    0
-#define L_NEW		1
-#define L_SEARCH	2
 
-
+void songdata_add_ordered( songdata *, songdata_song * );
+void songdata_del( songdata *, songdata_song * );
+void songdata_clear( songdata * );
+void songdata_add( songdata *, songdata_song *, songdata_song * );
 
 void songdata_read_mp3_list( songdata *, const char *, int );
 int	songdata_save_playlist( songdata *, char * );
 int songdata_check_file( songdata_song * );
 songdata_song * songdata_next_valid( songdata *, songdata_song *, int );
-void songdata_add( songdata *, songdata_song *, songdata_song * );
-void songdata_add_ordered( songdata *, songdata_song * );
-void songdata_del( songdata *, songdata_song * );
-void songdata_clear( songdata * );
-void dirstack_push( const char *, const char * );
-void dirstack_pop( void );
-char * dirstack_fullpath( void );
-char * dirstack_filename( void );
-int	dirstack_empty( void );
 
 songdata * songdata_init( void );
 void songdata_randomize( songdata * );
